Improve overflow check for type `int'
The C type `long' can be larger than the type `int'. Therefore, it is
verified that the value is >=INT_MIN and <=INT_MAX when the type is int.
Both option types `int' and `lint' use the C function `strtol' which
only performs overflow check for the `long' data type.