Add option type `enum'
This option type only accepts a predefined list of values.

This option type has a `values' property that defines a comma-separated
list of accepted values. For each enum option, an C enum is generated
that declares constants for its enum values.

The command-line option value is checked and converted into the
corresponding C enum value. The C type is `int'.