Add option property `replace'
This property defines a hash reference where all keys and values are
strings. When the option value is equal to a specific key, it is
replaced by the corresponding hash value. Thus, aliases for option
values can be defined.

The substitution is string-based and occurs before variable assignment.
Thus, type conversion occurs at runtime. However, this approach
simplifies the type system in GOPL. Additionally, this simplifies option
value verification etc.