Add support for optional option values
Add the option properties `optional' and `default'. If optional is set
to `yes', then the option takes an optional value. If a value is
specified in the same element of argv, that value is taken. Otherwise,
no value is given to that option.

When the `default' property is set, that value is used for variable
assignment as a fallback.

The option types can support optional values:
  - The type can be specified so that its value is optional. Then, a
	NULL string is used for variable assignment.
  - The type can require a value. When the option is set to `optional',
	then it must specify a `default' value that is used for the
	variable' assignment.

In summary, the following algorithm is used when the option takes an
optional value:
  1. When the same argv element contains a value, that one is used.
  2. When that argv element contains no value, the `default' property of
     the option is used. The default property must be specified unless
	 the option type allows that.
  3. When no value is specified at all (neither in the argv element nor
	 in the default property), then the value NULL is used for variable
	 assignment.

This commit also modifies the option types `string' and `callback' so
that it does not require an option value. For string options, the
get-function returns NULL at runtime.  However, the has-function returns
true to indicate that the option is given without value. For callback
options, the NULL value is passed to the callback function.