Create a DFA to parse option names
Use a deterministic finite automaton to parse the argument vector. The
DFA is generated in the following steps:
 1. Create a trie (prefix tree) containing all long names. This trie has
    compressed tails so that only the unique prefix is stored there.
 2. Create a DFA from the trie having the equivalent structure. This DFA
    matches the prefixes of option names
 3. Some parts of the DFA are generated manually, e.g. the part to match
    the leading double-dash.
 4. Short names use a special state in the DFA.

As the trie is a tree (which has no cycles), the DFA cannot have
equivalent states. Therefore, the DFA is minimal.

This approach leads to a new feature: The GNU calling convention says
that the option name does not need to be commplete. Instead, the option
name can be a prefix of an option as long as that prefix is not
ambigous. Therefore, the prefix-matching DFA seems to be an elegant
implementation. The non-ambigous tail of each option name is verified,
but this not done by the DFA. Instead, string operations are used here.

Only one variable assignment is generated for both long and short names.
This simplifies the code. The C variables use_short_name and option_name
name indicate which option name has been used.

Now, error messages show the correct option name. Previously, either the
short name or the first long name were used. When an abbreviation of a
long name has been used, the full name is displayed.