BUGFIX: Allow combination of short option names
This has been broken to detect wrong long options like --flag=val when
the flag option does not take a value. However, short options can have
additional characters in their argv element when they do not take a
value.

The Posix standard allows that multiple short options are combined in
one argv element. This can be done until one option takes a value, then
the rest of the argv element is taken as value of that option.