Use one buffer for short option names
Use a buffer on the stack for short option names. This simplifies the
code because option_name can point to that buffer now. The short option
names do not need to be a constsant string in the C code now.

Now, the C variable option_name is used to handle an invalid short
option name. Therefore, no unused... compiler warning will occur.