Improve help and version text format
Use two columns in the help text. The left column displays the
argument/option name and the right column contains the description.

All text is wrapped automatically so that it fits into the screen. The
screen size is defined statically using $config{pagewidth}, which
defaults to 80.

The language keys `help_desc' and `help_info' are removed because they
are unused in the new help text format. Both texts are printed using
automatic line wrapping. The first line has no indent, all other lines
are indented using $help{indent}.

The info field in the version text is modified similar to the info field
of the help text. In contrast, the indent is defined by $version{indent}
instead of $help{indent}.

The configuration property $config{pagewidth} is added which defines the
text width for all fields in both version and help texts. The default
value is 80.

The meaning if $help{indent2} is changed. It is now the indent used for
the second column at the option/argument description. The new default
value is " " x25.