Generate Makefile dependencies
Generate a dependency file using the -d switch. This option takes one
argument, which is the name of the dependency file.

The dependency file contains make rules for the -c and the -h files, as
long as those are given in @ARGV. The perl variable %INC is used for
dependencies.

Also clean up the getopt.pl script. The command line parsing is now done
in the "main" function.