Avoid unnatural loops
Split assignment states into different labels for long and short
options. Duplicate assignment states for non-argument options.

The assignment of short options without argument is an unnatural loop
which is hard for the compiler to optimize. The assignment of short
options is some kind of an inner loop over one argv element, whereas the
assignment of long options is an ounter loop over all argv elements.

Now, it is clear that it is impossible to jump from the outer loop into
the innter loop and vice versa. Therfore, the compiler can now easily
detect the inner and the outer loop. Otherwise, the compiler can hardly
apply optimization techniques to the opt_parse function.