Allow arguments to be specified in any order
Allow that options and non-option-arguments (NOAs) are given in any
order in argv. The GNU calling convention says that options and NOAs can
be specified in any order, except for the special argument --, which
disables option processing for the following arguments.

The opt_parse function now maintains an array which stores the type of
each argv element, which can either be option (stored in the array
`opts'), or NOA (stored in the array `args'). Here, option arguments are
treated like options, and the program name in argv[0] is ignored. The
double-dash argument marks all following argv elements as NOAs, but
is treated like the last option in argv, because it does not count as
NOA. After option processing, the opt_parse function permutes the argv
vector so that all options appear before all NOAs, with the double-dash
in between (if found in argv):

  progname options... ddash non-option-arguments...

After opt_parse, the argv vector has the form required by posix, where
all options must be given before all NOAs. The double-dash is
treated correct as well. It is neither option nor NOA, but it marks
the border of the argument types.

The order of options and NOAs remains unchanged, regarding arguments of
the same type. Only argv elements of different type are exchanged.

When the environment variable POSIXLY_CORRECT is set, the argv
permutation is disabled. Then, the first non-option element in argv
stops option processing and no permutation is performed.