Add support for option references
Flag options can be aliases for other options. They can assign special
values to the variables of other options. For instance, a `--quiet'
option can reset the `--verbose' variable to zero.

Two option properties are added:
  reference : declares one name of the referenced option
  value     : declares the value of the flag option, the default
              value is `1'. When a flag option is a reference, then the
			  value must be specified explicitly.

Only flag options support the reference and the value properties. When
the referenced option is a flag, then that option value is assigned to
the option variable. Otherwise, the option value is taken as a constant
C string, which is taken as option argument. Then, the standard variable
assignment process of the referenced variable is performed. For
instance, the verify function is called, if specified. Note that invalid
option values are not detected at compile time because variable
assignment is performed at runtime. This can lead to strange error
messages when a reference option tries to assign an invalid value.

The reference option does not provide a get-function. Therefore, it is
not visible in the C interface.