Generate struct member accessor macros
Declare macros to reliably access the struct members. The names are not
get/has as they have been before, because they can be used as
assignment target (then, the name *get* would be inprecise).

The macros exist because all option names should be valid in code, but
only one element in the struct exists for each option. Therefore, the
accessor macros can be used to reliably access the elements in the
struct, using any option name.

Also generate accessor macros for the configuration-independent struct
members.