Generate *_given fields for help and version types
Add _given members to the options struct. When using the break property,
it should be possible to find out which option caused the parser
termination. Therefore, these options get a _given member in the struct
and accessor macros. The function which invokes the opt_parse function
can then find out that this particular option triggered the end of
option processing.

When the exit property is set, then none of these options gets a _given
field because it is useless then.